+++ NEWS.md
@@ -7,6 +7,8 @@
 
 - `%in%.Splits()` retains names when comparing small splits
   ([#40](https://github.com/ms609/TreeTools/issues/40)).
+  
+- Improve handling of zero-split trees.
 
 - `DropTip()` no longer adds a root to unrooted trees.